Update runTestsAsyncPost to accept methods option
Updated the runTestsAsyncPost method to accept a methods option. If
passed, it will utilize the runTestsAsynchronous API method, send it the
JSON payload in the documentation. When passed, it will run only the
tests specified in the testMethods array (if present). You can then run
a whole test class or specific methods in a class.

Salesforce documentation:
https://developer.salesforce.com/docs/atlas.en-us.api_tooling.meta/api_tooling/intro_rest_resources.htm